Replace all outputs to usize

+++ 2022/rust/src/day1.rs
@@ -1,25 +1,25 @@
 use crate::parser;
 
-pub fn part1() -> u32 {
+pub fn part1() -> usize {
     let lines: Vec<String> = parser::read("data/day1.txt").unwrap();
     let food_calories = to_food_calories_per_elf(&lines);
     max_carried_calories(&food_calories)
 }
 
-pub fn part2() -> u32 {
+pub fn part2() -> usize {
     let lines: Vec<String> = parser::read("data/day1.txt").unwrap();
     let food_calories = to_food_calories_per_elf(&lines);
     max_top_3_carried_calories(&food_calories)
 }
 
-fn to_food_calories_per_elf(lines: &Vec<String>) -> Vec<Vec<u32>> {
+fn to_food_calories_per_elf(lines: &Vec<String>) -> Vec<Vec<usize>> {
     lines
         .split(|line| line.is_empty())
         .map(|lines| lines.iter().map(|s| s.parse().unwrap()).collect())
         .collect()
 }
 
-fn max_carried_calories(food_calories: &Vec<Vec<u32>>) -> u32 {
+fn max_carried_calories(food_calories: &Vec<Vec<usize>>) -> usize {
     food_calories
         .iter()
         .map(|calories| calories.iter().sum())
@@ -27,8 +27,8 @@ fn max_carried_calories(food_calories: &Vec<Vec<u32>>) -> u32 {
         .unwrap()
 }
 
-fn max_top_3_carried_calories(food_calories: &Vec<Vec<u32>>) -> u32 {
-    let mut total_calories: Vec<u32> = food_calories
+fn max_top_3_carried_calories(food_calories: &Vec<Vec<usize>>) -> usize {
+    let mut total_calories: Vec<usize> = food_calories
         .iter()
         .map(|calories| calories.iter().sum())
         .collect();
